+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.5.10
 
 * Fix company creation with identity upload and missing nostr node id during block propagation
+* rename `disable_mandatory_email_confirmations` to `mandatory_email_confirmations` with default = `true` (breaking CONFIG change)
 
 # 0.5.9